Extracted favorite button into its own component
