+++ src/app/projects/project-favorite/project-favorite-button.component.html
@@ -1,0 +1,3 @@
+<button (click)="onFavoriteIconClicked($event)"  aria-label="Toggle project as favorite">
+  <i [ngClass]="getClass()"></i>
+</button>
