+++ src/app/projects/project-list/project-list-item.component.ts
@@ -5,10 +5,6 @@ import { LocalStorageProjectsService } from '../local-storage-projects.service';
 import { ProjectRefreshedEvent } from '../project-refresh';
 import { ProjectUtilsService } from '../project-utils.service';
 
-export interface ProjectFavoriteClickEvent {
-  projectId: number;
-}
-
 @Component({
   selector: 'wh-project-list-item',
   templateUrl: './project-list-item.component.html',
@@ -17,7 +13,6 @@ export class ProjectListItemComponent implements OnChanges {
   @Input() project: WharfProject;
 
   @Output() refreshed = new EventEmitter<ProjectRefreshedEvent>();
-  @Output() favoriteClick = new EventEmitter<ProjectFavoriteClickEvent>();
 
   isRefreshAnimationPlaying: boolean;
   providerHostname: string;
@@ -35,11 +30,5 @@ export class ProjectListItemComponent implements OnChanges {
         this.providerHostname = this.project.provider.url;
       }
     }
-  }
-
-  onFavoriteIconClicked(event: MouseEvent) {
-    event.stopPropagation();
-    this.localStorageProjectsService.handleFavoriteButtonClick(this.project.projectId);
-    this.favoriteClick.emit({ projectId: this.project.projectId });
   }
 }
