+++ src/scss/imports.scss
@@ -14,5 +14,6 @@
 @import "nav.component.scss";
 @import "project-details.component.scss";
 @import "project-list-item.component.scss";
+@import "project-favorite-button.component.scss";
 @import "provider-form.component.scss";
 @import "provider.component.scss";
