+++ src/scss/project-details.component.scss
@@ -1,5 +1,3 @@
-@import "./button-icon-only.scss";
-
 .main-container {
   .project-page-header {
     p {
