+++ src/scss/project-list-item.component.scss
@@ -1,5 +1,3 @@
-@import "./button-icon-only.scss";
-
 wh-project-list-item {
   display: table-row;
 }
